Modify account to use an env variable
Switching accounts was a hassle with the account id embedded in the
script.  That was just laziness from the beginning, and now it's a
necessary use-case for me.

Added AWS_TOKEN_ACCOUNT variable and checking logic to make sure it's
set or at least properly updated in the function itself.